Move operator<< from VSIMEXMethod to DiscreteTime
