+++ applications/Christensen.cc
@@ -478,7 +478,8 @@ void Christensen<dim>::run()
     // Advances the VSIMEXMethod instance to t^{k}
     update_solution_vectors();
     time_stepping.advance_time();
-    *this->pcout << time_stepping << std::endl;
+    *this->pcout << static_cast<TimeDiscretization::DiscreteTime &>(time_stepping)
+                 << std::endl;
 
     // Performs post-processing
     if ((time_stepping.get_step_number() %
