+++ applications/MIT.cc
@@ -415,7 +415,8 @@ void MITBenchmark<dim>::run()
 {
   const unsigned int n_steps = this->prm.time_discretization_parameters.n_maximum_steps;
 
-  *this->pcout << time_stepping << std::endl;
+  *this->pcout << static_cast<TimeDiscretization::DiscreteTime &>(time_stepping)
+               << std::endl;
   while (time_stepping.get_current_time() < time_stepping.get_end_time() &&
          time_stepping.get_step_number() < n_steps)
   {
@@ -442,7 +443,8 @@ void MITBenchmark<dim>::run()
     // Advances the VSIMEXMethod instance to t^{k}
     update_solution_vectors();
     time_stepping.advance_time();
-    *this->pcout << time_stepping << std::endl;
+    *this->pcout << static_cast<TimeDiscretization::DiscreteTime &>(time_stepping)
+                 << std::endl;
 
     // Performs post-processing
     postprocessing();
